refactor(gui): extract navigation + first-launch into NavigationMixin
Move _switch_page (tail/info-auto-refresh wiring on page change),
_maybe_run_first_launch_checks (resume pending install + show welcome
dialog), and _show_quick_start (the welcome dialog) off WinpodxWindow
into a private mixin. These three methods form the first-run /
navigation flow; pairing them keeps the start-of-session UX a single
cohesive unit.

- NavigationMixin (src/winpodx/gui/_main_window_nav.py)
  _switch_page, _maybe_run_first_launch_checks, _show_quick_start.

WinpodxWindow now reads
`class WinpodxWindow(AppCrudMixin, HeaderMixin, InfoPageMixin,
LibraryPageMixin, LogsMixin, MaintenanceMixin, NavigationMixin,
PodStatusMixin, SettingsPageMixin, QMainWindow)`.

The .welcomed marker behaviour, the QTimer-deferred quick-start launch,
and the per-tab tail / auto-refresh ordering all come along as-is.
No behaviour change; tests: 1096 passed.

main_window.py: 243 -> 148 lines (-95). Cumulative: 2745 -> 148 (-95%).

WinpodxWindow is now a pure orchestrator: signal declarations,
__init__ state init, _setup_signals wiring, and _build_ui assembly.
All page builders, page behaviour, and chrome construction live in
single-responsibility mixin modules.
